Bug 1616591 - Make all the APZ log statements build again if enabled. r?botond
- Some breakage in FocusState/FocusTarget was introduced when the LayersId struct
  was introduced in bug 1448490; printing LayersId type now requires using the
  cast operator to get back the underlying uint64_t.
- Some breakage was introduced in APZCTreeManager by bug 1552608 when it
  introduced the HitTestResult structure.
- Some breakage in SimpleVelocityTracker seems to have been there since it
  landed as it tries to access a protected member in Axis. Added a public
  method to expose this as needed.
- Some breakage in ActiveElementManager was introduced in bug 1266595 when the
  runnable was wrapped in a RefPtr.

Differential Revision: https://phabricator.services.mozilla.com/D63547

+++ gfx/layers/apz/src/FocusState.cpp
@@ -65,8 +65,8 @@ void FocusState::Update(LayersId aRootLayerTreeId,
   MutexAutoLock lock(mMutex);
 
   FS_LOG("Update with rlt=%" PRIu64 ", olt=%" PRIu64 ", ft=(%s, %" PRIu64 ")\n",
-         aRootLayerTreeId, aOriginatingLayersId, aState.Type(),
-         aState.mSequenceNumber);
+         (uint64_t)aRootLayerTreeId, (uint64_t)aOriginatingLayersId,
+         aState.Type(), aState.mSequenceNumber);
   mReceivedUpdate = true;
 
   // Update the focus tree with the latest target
@@ -85,7 +85,7 @@ void FocusState::Update(LayersId aRootLayerTreeId,
     auto currentNode = mFocusTree.find(mFocusLayersId);
     if (currentNode == mFocusTree.end()) {
       FS_LOG("Setting target to nil (cannot find lt=%" PRIu64 ")\n",
-             mFocusLayersId);
+             (uint64_t)mFocusLayersId);
       return;
     }
 
@@ -130,11 +130,11 @@ void FocusState::Update(LayersId aRootLayerTreeId,
           FS_LOG(
               "Setting target to nil (bailing out of infinite loop, lt=%" PRIu64
               ")\n",
-              mFocusState.mFocusLayersId);
+              (uint64_t)mFocusState.mFocusLayersId);
           return true;
         }
 
-        FS_LOG("Looking for target in lt=%" PRIu64 "\n", aRefLayerId);
+        FS_LOG("Looking for target in lt=%" PRIu64 "\n", (uint64_t)aRefLayerId);
 
         // The focus target is in a child layer tree
         mFocusState.mFocusLayersId = aRefLayerId;
