+++ gfx/layers/apz/src/FocusTarget.cpp
@@ -164,7 +164,8 @@ FocusTarget::FocusTarget(PresShell* aRootPresShell,
     if (layersId.IsValid()) {
       FT_LOG("Creating reflayer target with seq=%" PRIu64 ", kl=%d, lt=%" PRIu64
              "\n",
-             aFocusSequenceNumber, mFocusHasKeyEventListeners, layersId);
+             aFocusSequenceNumber, mFocusHasKeyEventListeners,
+             (uint64_t)layersId);
 
       mData = AsVariant<LayersId>(std::move(layersId));
       return;
